fixed file size issue

+++ indexing/shard.c
@@ -2,11 +2,11 @@
 
 #define kCacheTableShardBufferSize 512
 
-static int _entry_compare(const unsigned char * b1, const unsigned char * b2, int len);
-static void _shard_base_insert_at(ShardNode * node, const unsigned char * data, int dataLength, int index);
-static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, int index);
+static long long _entry_compare(const unsigned char * b1, const unsigned char * b2, long long len);
+static void _shard_base_insert_at(ShardNode * node, const unsigned char * data, long long dataLength, long long index);
+static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, long long index);
 
-ShardNode * shard_node_new(int depthRem) {
+ShardNode * shard_node_new(long long depthRem) {
     ShardNode * node = (ShardNode *)malloc(sizeof(ShardNode));
     bzero((void *)node, sizeof(ShardNode));
     node->depthRemaining = depthRem;
@@ -20,7 +20,7 @@ void shard_node_free(ShardNode * node) {
             free(node->subnodes);
         }
     } else {
-        int i;
+        long long i;
         for (i = 0; i < node->subnodeCount; i++) {
             shard_node_free(node->subnodes[i]);
         }
@@ -34,15 +34,15 @@ void shard_node_free(ShardNode * node) {
 
 ShardNode * shard_node_search_base(ShardNode * root, 
                                    const unsigned char * data,
-                                   int entrySize, int create) {
+                                   long long entrySize, long long create) {
     if (entrySize == 0) return NULL;
     if (root->depthRemaining == 0) return root;
     // find the next subnode down using a binary search
-    int highIndex = root->subnodeCount;
-    int lowIndex = -1;
+    long long highIndex = root->subnodeCount;
+    long long lowIndex = -1;
     ShardNode * testNode = NULL;
     while (highIndex - lowIndex > 1) {
-        int testIndex = (highIndex + lowIndex) / 2;
+        long long testIndex = (highIndex + lowIndex) / 2;
         testNode = (ShardNode *)root->subnodes[testIndex];
         if (testNode->nodeCharacter == data[0]) {
             highIndex = testIndex;
@@ -54,7 +54,7 @@ ShardNode * shard_node_search_base(ShardNode * root,
             highIndex = testIndex;
         }
     }
-    int nodeIndex = (highIndex + lowIndex) / 2;
+    long long nodeIndex = (highIndex + lowIndex) / 2;
     if (nodeIndex >= 0 && nodeIndex < root->subnodeCount) {
         testNode = (ShardNode *)root->subnodes[nodeIndex];
         if (testNode->nodeCharacter == data[0]) {
@@ -77,8 +77,8 @@ ShardNode * shard_node_search_base(ShardNode * root,
 
 int shard_node_base_add(ShardNode * node, 
                         const unsigned char * data, 
-                        int entrySize, 
-                        int footerSize) {
+                        long long entrySize, 
+                        long long footerSize) {
     if (!node->nodeData) {
         node->nodeData = (unsigned char *)malloc(kCacheTableShardBufferSize * entrySize);
         node->nodeDataAlloc = kCacheTableShardBufferSize;
@@ -87,13 +87,13 @@ int shard_node_base_add(ShardNode * node,
         return 1;
     }
     // perform a binary search to find the insertion index
-    int lowestIndex = -1;
-    int highestIndex = node->nodeDataSize;
+    long long lowestIndex = -1;
+    long long highestIndex = node->nodeDataSize;
     while (highestIndex - lowestIndex > 1) {
-        int testIndex = (highestIndex + lowestIndex) / 2;
+        long long testIndex = (highestIndex + lowestIndex) / 2;
         const unsigned char * testObject = &node->nodeData[testIndex * entrySize];
-        int relativity = _entry_compare(testObject, data,
-                                       entrySize - footerSize);
+        long long relativity = _entry_compare(testObject, data,
+                                         entrySize - footerSize);
         if (relativity == 1) {
             highestIndex = testIndex;
         } else if (relativity == -1) {
@@ -104,12 +104,12 @@ int shard_node_base_add(ShardNode * node,
             break;
         }
     }
-    int insertIndex = (highestIndex + lowestIndex) / 2;
+    long long insertIndex = (highestIndex + lowestIndex) / 2;
     if (insertIndex < 0) insertIndex = 0;
     else if (insertIndex > node->nodeDataSize) insertIndex--;
     else {
         unsigned char * closeObject = &node->nodeData[insertIndex * entrySize];
-        int relativity = _entry_compare(closeObject, data, 
+        long long relativity = _entry_compare(closeObject, data, 
                                        entrySize - footerSize);
         if (relativity == 0) return 0;
         else if (relativity < 0) insertIndex++;
@@ -121,16 +121,16 @@ int shard_node_base_add(ShardNode * node,
 
 unsigned char * shard_node_base_lookup(ShardNode * node,
                                        const unsigned char * data,
-                                       int entrySize,
-                                       int footerSize) {
+                                       long long entrySize,
+                                       long long footerSize) {
     if (!node->nodeData || !node->nodeDataSize) return NULL;
-    int lowestIndex = -1;
-    int highestIndex = node->nodeDataSize;
+    long long lowestIndex = -1;
+    long long highestIndex = node->nodeDataSize;
     while (highestIndex - lowestIndex > 1) {
-        int testIndex = (highestIndex + lowestIndex) / 2;
+        long long testIndex = (highestIndex + lowestIndex) / 2;
         const unsigned char * testObject = &node->nodeData[testIndex * entrySize];
-        int relativity = _entry_compare(testObject, data,
-                                       entrySize - footerSize);
+        long long relativity = _entry_compare(testObject, data,
+                                        entrySize - footerSize);
         if (relativity == 1) {
             highestIndex = testIndex;
         } else if (relativity == -1) {
@@ -141,11 +141,11 @@ unsigned char * shard_node_base_lookup(ShardNode * node,
             break;
         }
     }
-    int bestIndex = (highestIndex + lowestIndex) / 2;
+    long long bestIndex = (highestIndex + lowestIndex) / 2;
     if (bestIndex < 0) return NULL;
     if (bestIndex >= node->nodeDataSize) return NULL;
     unsigned char * closeObject = &node->nodeData[bestIndex * entrySize];
-    int relativity = _entry_compare(closeObject, data, 
+    long long relativity = _entry_compare(closeObject, data, 
                                     entrySize - footerSize);
     if (relativity == 0) return closeObject;
     return NULL;
@@ -157,23 +157,23 @@ unsigned char * shard_node_base_lookup(ShardNode * node,
 
 static void _shard_base_insert_at(ShardNode * node, 
                                   const unsigned char * data, 
-                                  int dataLength, int index) {
+                                  long long dataLength, long long index) {
     if (node->nodeDataSize >= node->nodeDataAlloc) {
         node->nodeDataAlloc += kCacheTableShardBufferSize;
         node->nodeData = realloc(node->nodeData, node->nodeDataAlloc * dataLength);
     }
-    int copyCount = node->nodeDataSize - index;
+    long long copyCount = node->nodeDataSize - index;
     if (copyCount > 0) {
         unsigned char * memStart = &node->nodeData[index * dataLength];
         unsigned char * memDest = &node->nodeData[(index + 1) * dataLength];
-        int memSize = copyCount * dataLength;
+        long long memSize = copyCount * dataLength;
         memmove(memDest, memStart, memSize);
     }
     memcpy(&node->nodeData[index * dataLength], data, dataLength);
     node->nodeDataSize++;
 }
 
-static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, int index) {
+static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, long long index) {
     if (!node->subnodes) {
         node->subnodes = malloc(sizeof(void *) * 8);
         node->subnodeAlloc = 8;
@@ -181,11 +181,11 @@ static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, in
         node->subnodeAlloc += 8;
         node->subnodes = realloc(node->subnodes, sizeof(void *) * node->subnodeAlloc);
     }
-    int moveCount = node->subnodeCount - index;
+    long long moveCount = node->subnodeCount - index;
     if (moveCount > 0) {
         void * memStart = (void *)&node->subnodes[index];
         void * memDest = (void *)&node->subnodes[index + 1];
-        int memSize = moveCount * sizeof(void *);
+        long long memSize = moveCount * sizeof(void *);
         memmove(memDest, memStart, memSize);
     }
     node->subnodes[index] = subnode;
@@ -193,8 +193,8 @@ static void _shard_node_insert_subnode(ShardNode * node, ShardNode * subnode, in
 }
 
 
-static int _entry_compare(const unsigned char * b1, const unsigned char * b2, int len) {
-    int i;
+static long long _entry_compare(const unsigned char * b1, const unsigned char * b2, long long len) {
+    long long i;
     for (i = 0; i < len; i++) {
         if (b1[i] > b2[i]) return 1;
         if (b1[i] < b2[i]) return -1;
